adding a new test for ceph self recovery
